refactor(intent): delegate diary intents to DiaryService
