+++ Cookle/Sources/Main/Services/MainService.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+@MainActor
+enum MainService {
+    static func open() throws {}
+}
+
