+++ Cookle/Sources/Recipe/Intents/SearchRecipesIntent.swift
@@ -23,19 +23,7 @@ struct SearchRecipesIntent: AppIntent, IntentPerformer {
     @Dependency private var modelContainer: ModelContainer
 
     static func perform(_ input: Input) throws -> Output {
-        let (context, text) = input
-        var recipes = try context.fetch(
-            .recipes(.nameContains(text))
-        )
-        let ingredients = try context.fetch(
-            text.count < 3 ? .ingredients(.valueIs(text)) : .ingredients(.valueContains(text))
-        )
-        let categories = try context.fetch(
-            text.count < 3 ? .categories(.valueIs(text)) : .categories(.valueContains(text))
-        )
-        recipes += ingredients.flatMap(\.recipes.orEmpty)
-        recipes += categories.flatMap(\.recipes.orEmpty)
-        return Array(Set(recipes))
+        try RecipeService.search(context: input.context, text: input.searchText)
     }
 
     func perform() throws -> some ReturnsValue<[RecipeEntity]> {
